rename util service
