--- src/util.service.js
@@ -1,17 +1,0 @@
-export default class UtilService {
-  static modeEmailHtml = 'email-html';
-
-  static modeEmailMjml = 'email-mjml';
-
-  static modePageHtml = 'page-html';
-
-  static getMode(editor) {
-    const cfg = editor.getConfig();
-
-    if (!cfg.pluginsOpts || !cfg.pluginsOpts.grapesjsmautic) {
-      throw new Error('Wrong Mautic Grapesjs mode');
-    }
-
-    return cfg.pluginsOpts.grapesjsmautic.mode;
-  }
-}
